did some refactoring of the navbar
