--- src/App/components/Navigation-Build/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Navigation-Build';
