--- src/App/components/Navigation-Public/Navigation-Public.module.scss
@@ -1,48 +1,0 @@
-@import '../../../index.scss';
-
-.user-welcome {
-  padding: 5px;
-  margin-right: 15px;
-  text-align: right;
-}
-
-.user-welcome:hover {
-  color: $pure-yellow;
-  transition: 0.5s;
-}
-
-.navbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
-
-.button_things,
-.user_things {
-  display: none;
-}
-
-.mobile-button {
-  display: flex;
-  flex-direction: column;
-  margin: 10px;
-  padding: 10px;
-}
-
-.home-title {
-  background-image: url('../../Images/PokeTeams---Home.png');
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: 90%;
-  height: 12em;
-  margin: auto;
-}
-
-@media (min-width: 850px) {
-  /* Large Ipad Screen */
-
-  .home-title {
-    background-size: 70%;
-    height: 15em;
-  }
-}
