--- src/App/components/Navigation-Public/Navigation-Public.test.tsx
@@ -1,26 +1,0 @@
-import React, { Suspense } from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-//import { render } from '@testing-library/react';
-import NavigationPublic from './Navigation-Public';
-import { shallow, mount, render } from 'enzyme';
-import toJson from 'enzyme-to-json';
-
-describe('<NavigationPublic/>', (): void => {
-  it('renders without crashing', (): void => {
-    const div: HTMLDivElement = document.createElement('div');
-    ReactDOM.render(
-      <BrowserRouter>
-        <Suspense fallback={<div>test</div>}>
-          <NavigationPublic />
-        </Suspense>
-      </BrowserRouter>,
-      div
-    );
-    ReactDOM.unmountComponentAtNode(div);
-  });
-  it('snapshot', (): void => {
-    const wrapper = shallow(<NavigationPublic />);
-    expect(toJson(wrapper)).toMatchSnapshot();
-  });
-});
