--- src/App/components/Navigation-Public/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Navigation-Public';
