--- src/App/components/SearchBar-Build/SearchBar-Build.module.scss
@@ -1,34 +1,0 @@
-@import '../../../index.scss';
-
-.searchbar,
-.search {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: auto;
-}
-
-.s-input {
-  font-family: $body-font;
-  border: none;
-  border-radius: 4px;
-}
-
-@media (min-width: 700px) {
-  /* Medium Screen */
-  .searchbar,
-  .search {
-    justify-content: space-between;
-    align-items: center;
-    flex-direction: row;
-  }
-
-  .button_things,
-  .user_things {
-    display: inline;
-  }
-
-  .mobile-button {
-    display: none;
-  }
-}
