--- src/App/components/SearchBar-Build/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchBar-Build';
