--- src/App/components/SearchBar-Public/SearchBar-Public.module.scss
@@ -1,42 +1,0 @@
-@import '../../../index.scss';
-
-.searchbar,
-.search {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: auto;
-}
-
-.search {
-  margin-top: 5%;
-}
-
-.s-input {
-  font-family: $body-font;
-  border: none;
-  border-radius: 4px;
-}
-
-@media (min-width: 700px) {
-  /* Medium Screen */
-  .searchbar,
-  .search {
-    justify-content: space-between;
-    align-items: center;
-    flex-direction: row;
-  }
-
-  .search {
-    margin-top: 12%;
-  }
-
-  .button_things,
-  .user_things {
-    display: inline;
-  }
-
-  .mobile-button {
-    display: none;
-  }
-}
