--- src/App/components/SearchBar-Public/SearchBar-Public.test.tsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-//import { render } from '@testing-library/react';
-import SearchBarPublic from './SearchBar-Public';
-import { shallow, mount, render } from 'enzyme';
-import toJson from 'enzyme-to-json';
-
-describe('<SearchBarPublic/>', (): void => {
-  it('renders without crashing', (): void => {
-    const div: HTMLDivElement = document.createElement('div');
-    ReactDOM.render(
-      <BrowserRouter>
-        <SearchBarPublic />
-      </BrowserRouter>,
-      div
-    );
-    ReactDOM.unmountComponentAtNode(div);
-  });
-  it('snapshot', (): void => {
-    const wrapper = shallow(<SearchBarPublic />);
-    expect(toJson(wrapper)).toMatchSnapshot();
-  });
-});
