--- src/App/components/SearchBar-Public/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchBar-Public';
