+++ src/App/components/SearchBar/SearchBar.module.scss
@@ -1,0 +1,79 @@
+@import '../../../index.scss';
+
+.searchbar,
+.search {
+  display: flex;
+  flex-direction: column;
+  margin: auto;
+  height: 7.8125rem;
+}
+
+.search {
+  justify-content: space-around;
+  margin-top: 5%;
+}
+
+.label-font {
+  font-family: $h2-font;
+  color: $super-light-cream;
+  font-size: 1rem;
+}
+
+.select {
+  display: flex;
+  flex-direction: row;
+  label,
+  select {
+    font-size: 1rem;
+  }
+  label {
+    font-family: $h2-font;
+    color: $super-light-cream;
+  }
+}
+
+.s-input {
+  font-family: $body-font;
+  font-size: 1rem;
+  border: none;
+  border-radius: 0.25rem;
+}
+
+@media (min-width: 700px) {
+  /* Medium Screen */
+
+  .button {
+    font-size: 1.3rem;
+  }
+  .searchbar,
+  .search {
+    justify-content: center;
+    align-items: center;
+    flex-direction: row;
+    margin: 0;
+    height: unset;
+  }
+
+  .button_things,
+  .user_things {
+    display: inline;
+  }
+
+  .mobile-button {
+    display: none;
+  }
+  .label-font {
+    font-size: 1.3rem;
+  }
+  .select {
+    display: flex;
+    flex-direction: row;
+    label,
+    select {
+      font-size: 1.3rem;
+    }
+  }
+  .s-input {
+    font-size: 1.3rem;
+  }
+}
