+++ src/App/routes/BuildPage/BuildPage.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment, FunctionComponent } from 'react';
-import NavigationBuild from '../../components/Navigation-Build/Navigation-Build';
+import Navigation from '../../components/Navigation';
 import AddTeamSection from '../../components/AddTeamSection/AddTeamSection';
 import TeamListEdit from '../../components/TeamList-Edit/TeamList-Edit';
 import FolderList from '../../components/FoldersList/FoldersList';
@@ -8,7 +8,7 @@ import Footer from '../../components/Footer/Footer';
 const BuildPage: FunctionComponent = (): JSX.Element => {
   return (
     <Fragment>
-      <NavigationBuild />
+      <Navigation isPublic={false} />
       <main>
         <header role="banner">
           <FolderList />
