+++ src/App/routes/HomePage/HomePage.tsx
@@ -1,12 +1,12 @@
 import React, { Fragment, FunctionComponent } from 'react';
-import NavigationPublic from '../../components/Navigation-Public/Navigation-Public';
+import Navigation from '../../components/Navigation';
 import TeamListPublic from '../../components/TeamList-Public/TeamList-Public';
 import Footer from '../../components/Footer/Footer';
 
 const HomePage: FunctionComponent = (): JSX.Element => {
   return (
     <Fragment>
-      <NavigationPublic />
+      <Navigation isPublic />
       <main>
         <header role="banner">
           <h2>Results:</h2>
